Move connection_db to a new file and start adding tests
